delete gem interactor

+++ app/interactors/books/delete.rb
@@ -1,14 +1,13 @@
 module Books
   class Delete
-    include Interactor
-    
-    def call
-      book = ::Book.find_by(id: context.id)
-      context.fail!(error: "Book not found") unless book
+    def self.call(id:)
+      book = ::Book.find_by(id: id)
+      return { success: false, error: "Book not found" } unless book
+
       if book.destroy
-        context.book = book
+        { success: true, book: book }
       else
-        context.fail!(error: book.errors.full_messages)
+        { success: false, error: book.errors.full_messages }
       end
     end
   end
